pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ikdd-runtime"
version = "0.2.0"
description = "IKDD Hybrid AI Runtime"
requires-python = ">=3.10"

# Dependencies to pip install if needed for runtime
dependencies = [
    "pyyaml", "openai", "anthropic", "requests"
]

# ★ CLI entry point ★
[project.scripts]
ikdd = "runtime.v0_2.cli:main"
